requirements.txt
autopep8
bokeh
click
colorama
face-recognition
face-recognition-models
future
imageio
importlib-metadata
Jinja2
MarkupSafe
numpy
opencv-contrib-python
opencv-python-headless
packaging
Pillow
pycodestyle
pyee
pyparsing
ffmpeg-python
PyYAML
timecode
toml
tornado
typing_extensions
wincertstore
zipp
odfpy